Creando .c para sysfs y cambiando nombre de archivo principal 	deleted:    FanCtlModule.c 	modified:   Makefile 	new file:   fanCtl.c 	new file:   sysfs.c 	new file:   sysfs.h
