+++ FanCtlModule/Makefile
@@ -1,5 +1,7 @@
 obj-m += FanCtlModule.o
 
+FanCtlModule-y += fanCtl.o sysfs.o
+
 all:
 	make -C /lib/modules/$(shell uname -r)/build M=$(PWD) modules
