+++ FanCtlModule/sysfs.h
@@ -1,0 +1,5 @@
+
+int crea_arbol_sysfs(void);
+
+int destruye_arbol_sysfs(void);
+
